review comments
other_s3_bucket_name -> api_s3_bucket_name

transforming data taken out into utility function
